LABDSOF-146 Added Verifications and Points Transaction

+++ netquest-backend/src/main/java/com/netquest/domain/pointsearntransaction/dto/PointsEarnTransactionDto.java
@@ -6,6 +6,6 @@ import java.util.UUID;
 public record PointsEarnTransactionDto(
         UUID id,
         LocalDateTime datetime,
-        int amount
+        long amount
 ) {
 }
